[Feat]진화용 카메라 추가

+++ Assets/Scripts/Manager/IHIt.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHIt 
 {
-    void Hit(int damage);
+    void Hit(float damage);
 }
